MathQuestionRepository composed
