--- src/main/java/pro/sky/java/course2/JavaQuestionsRepository.java
@@ -1,105 +1,0 @@
-package pro.sky.java.course2;
-
-import org.springframework.stereotype.Service;
-
-import javax.annotation.PostConstruct;
-import java.util.Collection;
-import java.util.HashSet;
-
-@Service
-public class JavaQuestionsRepository implements QuestionRepository {
-
-    private final Collection<Question> questionCollection = new HashSet<>();
-
-    JavaQuestionsRepository() {
-    }
-
-    @Override
-    public Question add(Question question) {
-        checkQuestion(question);
-        questionCollection.add(question);
-        return question;
-    }
-
-    private void checkQuestion(Question question) {
-        if (question == null || questionCollection.contains(question)) {
-            throw new QuestionServiceException("question is null or already added");
-        }
-    }
-
-
-    @Override
-    public Question add(String problem, String answer) {
-        Question question = new Question(problem, answer);
-        return add(question);
-    }
-
-    @Override
-    public Question remove(Question question) {
-        if (!questionCollection.contains(question)) {
-            throw new QuestionServiceException("question is missing");
-        }
-        questionCollection.remove(question);
-        return question;
-    }
-
-    @Override
-    public Question remove(String problem, String answer) {
-        Question question = new Question(problem, answer);
-        return remove(question);
-    }
-
-    @Override
-    public Collection<Question> getAll() {
-        Collection<Question> questions = new HashSet<>();
-        questions = questionCollection;
-        return questions;
-    }
-
-    @PostConstruct
-    public void loadJavaQuestions() {
-        add("What is Autoboxing and Unboxing?",
-                "Autoboxing is the automatic conversion that the Java compiler makes between the primitive types and their corresponding object wrapper classes. For example, converting an int to an Integer, a double to a Double, and so on. If the conversion goes the other way, this is called unboxing");
-        add("What class in Java is parent of Object class?",
-                "Object class is topmost class in Java, so it has no parent class");
-        add("What are the supported platforms by Java Programming Language?",
-                "Java runs on a variety of platforms, such as Windows, Mac OS, and the various versions of UNIX/Linux like HP-Unix, Sun Solaris, Redhat Linux, Ubuntu, CentOS, etc. This Particular section is dedicated to Question & Answer only. If you want learn more about Java Programming Language");
-        add("What are the differences between Heap and Stack Memory in Java?",
-                "Stack memory in data structures is the amount of memory allocated to each individual programme. It is a fixed memory space. Heap memory, in contrast, is the portion that was not assigned to the Java code but will be available for use by the Java code when it is required, which is generally during the program's runtime.");
-        add("How does Java enable high performance?",
-                "Java uses Just In Time compiler to enable high performance. It is used to convert the instructions into bytecodes");
-        add("What do you mean by Constructor?",
-                "Constructors can be explained in detail with enlisted points:\n" +
-                        "When a new object is created in a program a constructor gets invoked corresponding to the class.\n" +
-                        "The constructor is a method which has the same name as the class name.\n" +
-                        "If a user doesn’t create a constructor implicitly a default constructor will be created.\n" +
-                        "The constructor can be overloaded.\n" +
-                        "If the user created a constructor with a parameter then he should create another constructor explicitly without a parameter"
-        );
-
-        add("What is a Class?",
-                "All Java codes are defined in a Class. It has variables and methods.\n" +
-                        "Variables are attributes which define the state of a class.\n" +
-                        "Methods is a place where the exact business logic has to be done. It contains a set of statements (or) instructions to satisfy the particular requirement.");
-        add("What is meant by the Local variable and the Instance variable?",
-                "Local variables are defined in the method and scope of the variables that exist inside the method itself. Instance variable is defined inside the class and outside the method and the scope of the variables exists throughout the class.");
-        add("What is an Object?",
-                "An instance of a class is called an object. The object has state and behavior. Whenever the JVM reads the “new()” keyword then it will create an instance of that class");
-        add("What is Inheritance?", "Inheritance means one class can extend to another class. So that the codes can be reused from one class to another class. The existing class is known as the Super class whereas the derived class is known as a sub class");
-        add("What is Encapsulation?",
-                "Purpose of Encapsulation:\n" + "Protects the code from others.\n" + "Code maintainability");
-        add("What is meant by Method Overriding?",
-                "Method overriding happens if the sub-class method satisfies the below conditions with the Super-class method:\n" +
-                        "Method name should be the same\n" +
-                        "The argument should be the same\n" +
-                        "Return type should also be the same\n" +
-                        "The key benefit of overriding is that the Sub-class can provide some specific information about that sub-class type than the super-class");
-        add("What is meant by Overloading?", "Method overloading happens for different classes or within the same class.\n" +
-                "For method overloading, sub-class method should satisfy the below conditions with the Super-class method (or) methods in the same class itself:\n" +
-                "Same method name\n" +
-                "Different argument types\n" +
-                "There may be different return types");
-
-    }
-
-}
